Converted singleChoice to true/false. Allowed editing interaction response options on editSlideForm. Added correct / incorrect response option indicator on editSlideForm.  Added interaction card showing options, responses and toggle button for content slide.
