+++ src/interaction/components/host/Content.vue
@@ -139,6 +139,33 @@ onUpdated(() => {
       </ul>
     </div>
   </div>
+  <div
+    v-if="slide.isInteractive"
+    class="d-flex flex-wrap align-items-center align-content-stretch justify-content-around"
+  >
+    <div class="card bg-transparent shadow p-2 m-2 align-items-center">
+      <h4>Interaction</h4>
+      <div class="d-flex flex-wrap mx-2">
+        <div>
+          Options:
+          <ul>
+            <li v-for="option in slide.interaction.options" :key="option">
+              {{ option }}
+            </li>
+          </ul>
+        </div>
+        <div class="d-flex flex-column align-items-center mx-2">
+          Responses: {{ slide.interaction.submissions.length }}
+          <button
+            class="btn btn-teal btn-sm m-4"
+            @click="$emit('toggleContent')"
+          >
+            Show responses
+          </button>
+        </div>
+      </div>
+    </div>
+  </div>
 </template>
 
 <style scoped>
